refactor: dto 패키지 계층 및 이너 클래스 구조 분리
