+++ backend/src/main/java/woowacourse/touroot/authentication/controller/LoginController.java
@@ -13,7 +13,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import woowacourse.touroot.authentication.dto.LoginResponse;
+import woowacourse.touroot.authentication.dto.response.LoginResponse;
 import woowacourse.touroot.authentication.service.LoginService;
 import woowacourse.touroot.global.exception.dto.ExceptionResponse;
