+++ backend/src/main/java/woowacourse/touroot/authentication/dto/response/kakao/KakaoAccount.java
@@ -1,0 +1,6 @@
+package woowacourse.touroot.authentication.dto.response.kakao;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public record KakaoAccount(@JsonProperty("profile") KakaoProfile kakaoProfile) {
+}
